Suppression de livre et début du Power Point

+++ back-end/controllers/books.js
@@ -1,3 +1,4 @@
+const fs = require('fs');
 const Book = require('../models/Book');
 
 //Capture et enregistre l'image, analyse le livre transformé en chaîne de caractères, et l'enregistre dans la base de données en définissant
@@ -5,14 +6,17 @@ const Book = require('../models/Book');
 //demande initiale est vide ; lorsque Multer est ajouté, il renvoie une chaîne pour le corps de la demande en fonction des données soumises avec le fichier.
 exports.createBook = (req, res) => 
 {
-    delete req.body._id;
-    console.log(req.body);
+    const bookObject = JSON.parse(req.body.book);
+    delete bookObject._id;
+    delete bookObject._userId;
     const book = new Book({
-        ...req.body
+        ...bookObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     book.save()
-        .then(() => res.status(201).json({ message: 'Livre enregistré !'}))
-        .catch(error => res.status(400).json({ error }));
+    .then(() => { res.status(201).json({message: 'Livre ajouté !'})})
+    .catch(error => { res.status(400).json( { error })})
 };
 
 //Définit la note pour le user ID fourni. La note doit être comprise entre 0 et 5. L'ID de l'utilisateur et la note doivent être ajoutés au
@@ -21,36 +25,71 @@ exports.createBook = (req, res) =>
 //jour, et le livre renvoyé en réponse de la requête.
 exports.rateBook = (req, res) => 
 {
-    delete req.body._id;
-    const book = new Book({
-        ...req.body
-    });
-    book.save()
-        .then(() => res.status(201).json({ message: 'Livre enregistré !'}))
+    
+    Book.findOne({ _id: req.params.id})
+    .then((book) => {
+        book.ratings.push({grade: req.body.rating , userId: req.body.userId})
+        var somme = 0;
+        book.ratings.map((rating) =>
+            somme = somme + rating.grade
+        )
+        var moyenne = somme / book.ratings.length;
+        book.averageRating = moyenne;
+        book.save()
+        .then(() => res.status(201).json(book))
         .catch(error => res.status(400).json({ error }));
+    })
 }
 
 //Met à jour le livre avec l'_id fourni
 exports.modifyBook = (req, res) => 
 {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Livre modifié !'}))
-        .catch(error => res.status(400).json({ error }));
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id})
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message : 'Not authorized'});
+            } else {
+                Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
+                .then(() => res.status(200).json({message : 'Livre modifié!'}))
+                .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
 };
 
 //Supprime le livre avec l'_id fourni ainsi que l’image associée.
 exports.deleteBook = (req, res) => 
 {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Livre supprimé !'}))
-        .catch(error => res.status(400).json({ error }));
+    Book.findOne({ _id: req.params.id})
+    .then(book => {
+        if (book.userId != req.auth.userId) {
+            res.status(401).json({message: 'Not authorized'});
+        } else {
+            const filename = book.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Book.deleteOne({_id: req.params.id})
+                    .then(() => { res.status(200).json({message: 'Livre supprimé !'})})
+                    .catch(error => res.status(401).json({ error }));
+            });
+        }
+    })
+    .catch( error => {
+        res.status(500).json({ error });
+    });
 }
 
 //Renvoie le livre avec l’_id fourni.
 exports.getOneBook = (req, res) => 
 {
-    Thing.findOne({ _id: req.params.id })
-        .then(thing => res.status(200).json(thing))
+    Book.findOne({ _id: req.params.id })
+        .then(book => res.status(200).json(book))
         .catch(error => res.status(404).json({ error }));
 }
 
@@ -58,14 +97,14 @@ exports.getOneBook = (req, res) =>
 exports.getAllBooks =  (req, res) => 
 {
     Book.find()
-        .then(things => res.status(200).json(things))
+        .then(books => res.status(200).json(books))
         .catch(error => res.status(400).json({ error }));
 }
 
 //Renvoie un tableau des 3 livres de la base de données ayant la meilleure note moyenne.
 exports.getBestRatedBooks =  (req, res) => 
 {
-    Thing.findOne({ _id: req.params.id })
-        .then(thing => res.status(200).json(thing))
+    Book.find().sort({'averageRating': -1}).limit(3)
+        .then(book => res.status(200).json(book))
         .catch(error => res.status(404).json({ error }));
 }
